requirements.txt
python-zenity
pyserial